apps/llamaindex-py/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
qdrant-client>=1.10,<2.0

# LlamaIndex core + integrations (namespaced packages)
llama-index-core>=0.12,<0.13
llama-index-vector-stores-qdrant>=0.4,<1.0
llama-index-llms-openai>=0.4,<1.0
llama-index-embeddings-openai>=0.5,<1.0
llama-index-llms-ollama>=0.5,<1.0
llama-index-embeddings-ollama>=0.3,<1.0

# Parsing
pypdf>=4.2,<5.0
python-docx>=1.1,<2.0

# Observability (OpenTelemetry)
opentelemetry-api>=1.26,<2.0
opentelemetry-sdk>=1.26,<2.0
opentelemetry-exporter-otlp>=1.26,<2.0
opentelemetry-instrumentation-fastapi>=0.47b0,<0.60
opentelemetry-instrumentation-requests>=0.47b0,<0.60
opentelemetry-instrumentation-logging>=0.47b0,<0.60
